Time: 52 ms (96.34%), Space: 15.6 MB (95.26%) - LeetHub

+++ 187-repeated-dna-sequences/187-repeated-dna-sequences.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     vector<string> findRepeatedDnaSequences(string s) {
-        unordered_map<long long, int> hasBefore;
+        unordered_map<int, int> hasBefore;
         vector<string> v;
         for(int i=0; i<=(int)s.size()-10; i++) {
-            long long ans = 0;
+            int ans = 0;
             for(int j=i; j<i+10; j++) {
                 ans = (5 * ans + (s[j]-'A'+1)%5);
             }
